⚗️  (cache) turbo check `main|[branch]|[prs]` (ʘ‿ʘ) (#358)
- [x] turbo pulls from latest cache and re-checks if package has been updated
  - Verification:  `@jeromefitz/semantic-config`
  - ⚗️ (semantic-config) break turbo cache for semantic-config
- [x] turbo pulls from latest cache and re-checks if multiple packages have been updated
  - Verification: `@jeromefitz/semantic-config|@jeromefitz/design-system`
  - 🔥 (semantic-config) remove .gitkeep
  - 🩹 (design-system) fix swr peer dependency
    - Please note this will get squashed and updated in next actual release for `@jeromefitz/design-system` 


Now in "theory" this when it gets merged to `main`:

- Grab the latest `node|yarn` cache based off of `yarn.lock`
- Grab the latest `turbo` by check:
  - `turbo-${{ github.job }}-${{ github.ref_name }}-${{ github.sha }}`
  - `turbo-${{ github.job }}-${{ github.ref_name }}-`
  - `turbo-${{ github.job }}-`
  - `turbo-`

Latest key should grab from `turbo-` then will check against any code changes that have come in since.

Now ... I say, or write, "theory" because I am now thinking even if this does work as intended what happens when this gets moved to the `[redacted]` ecosystem when we have 100s of developers working doing PRs from their forks (that should be fine) into "feature" or prerelease branches by build engineers?

Will that cause unintended consequences where the latest `turbo-` may be from a PR that was generated?

I think `turbo` will do its double-check and just run again on the package itself.

This entire repo is a proof of concept that I port over to a few places, haha. So will report back in a few months on scale (or weeks).