🔧  (renovate) more tweaks
renovate started doubling up the commit type/scope 🎲

determine if we can have more granular control over
- dependencies
- devDependencies
- peerDependencies

Ideally, in packages that are to be released (!root),
 we want to unlock down some of the `dev|peer` constraints

By that, I think we want to move quickly with them,
 but we do not need to semver release on dev|peer updates.

They are "safe" to be held until the next:
 - feature|minor
 - fix|patch

As they deal with compilation. May be ideal for a deeper
 dive on use of the packages by downstream codebases.

Are we being too stringent on the pinning? How can
 we be good stewards with lock files and install sizes.

Can we rely _more_ on `peerDependencies` and less on direct?

Obviously if anyone is reading this, you may see an initiative
 or two that is being worked on at a larger scale than this
 repo, haha.

🔧  (renovate) more tweaks

🔧  (renovate) no semantic via renovate, since we juse gitmoji

This is causing confusion all-over.
For now, just _set_ the `commitMessagePrefix` the way
 we want throughout without jamming this into existing
 presets.

🔧  (renovate) seems to always detect 🤔

Preference _would_ be to use semantic.
Goal would be to get it in a quickly usable state
 of our current flow to determine the `dep-dev|peer` decision

Then coming back at a later date to determine how to use
 renovate as intended.

🔧  (renovate) valhalla, some combination has to eventually work

🔧  (renovate) !commitPrefix

Not going to win this battle with renovate, haha.