✨  (ccommit) enquirer ➖ inquirer|chalk [b] (#983)
- [x] Migrate to `enquirer`
- [x] Update to follow `enquirer` log format
- [x] Prep for `breaking change` in subject but comment out
- [x] Prep for `pkg` but separate that out into a future PR
- [x] Add `.js` and `/index.js` where appropriate
- [x] Add `node:xyz` imports where appropriate  


Pretty one-for-one swap here, and there is no breaking change. A lot of code pruning as a result.

This opens up the roadmap a lot easier. Some tradeoffs on the Character Counter at the moment with `validate|format` but more than willing to take that on later.

Excited to introduce configuration files / cache to speed things up all-around and allow people to take more control over this. Probably should have kept this prerelease, but here we are.